\name{Diabetes}
\alias{Diabetes}
\title{Diabetes}
\docType{data}
\description{
This diabetes data set has n = 352 samples and there are p = 6 predictors: lamotrigine (ltg), total serum cholesterol (tc), total cholesterol (tch), low- and high-density lipoprotein (ldl and hdl) and glucose (glu). The response variable is the measurement of the disease progression one year after baseline.
}
\references{
B. Efron, T. Hastie, I. Johnstone, and R. Tibshirani. Least angle regression. Annals of statistics, 32(2):407–499, 2004.
}
\usage{Diabetes}
\format{A dataframe containing 352 records}
